backend/api/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
pydantic>=2.8,<3
supabase>=2.4,<3
boto3>=1.34,<2
python-dotenv>=1.0,<2
